:recycle: [#2471] merge appointment tasks and services
