+++ src/openforms/appointments/tasks.py
@@ -6,7 +6,8 @@ from openforms.celery import app
 from openforms.submissions.models import Submission
 
 from .exceptions import AppointmentRegistrationFailed
-from .service import register_appointment
+from .models import AppointmentInfo
+from .utils import book_appointment_for_submission
 
 __all__ = ["maybe_register_appointment"]
 
@@ -31,8 +32,21 @@ def maybe_register_appointment(submission_id: int) -> None:
     """
     logger.info("Registering appointment for submission %d (if needed!)", submission_id)
     submission = Submission.objects.get(id=submission_id)
+
     try:
-        register_appointment(submission)
+        appointment_id = submission.appointment_info.appointment_id
+    except AppointmentInfo.DoesNotExist:
+        pass
+    else:
+        # idempotency - do not register a new appointment if there already is one.
+        if appointment_id:
+            logger.info(
+                "Submission %s already has an appointment ID, aborting.", submission.pk
+            )
+            return
+
+    try:
+        book_appointment_for_submission(submission)
     except AppointmentRegistrationFailed as exc:
         logger.info(
             "Appointment registration failed, aborting workflow.",
